requirements.txt
# Definition of the Python Virtual Environment.
#
# WARNING: Only the required packages are in the environment.
#          The following optional libraries are not installed:
#              - FFTW, UMFPACK, and MKL/PARDISO
#              - CuPy and CUDA
#              - Jupyter
#
# Thomas Guillod - Dartmouth College
# Mozilla Public License Version 2.0

# YAML parsing library
PyYAML>=6.0

# numerical libraries
numpy>=1.24
scipy>=1.9
joblib>=1.2

# Qt bindings
PyQt5>=5.15

# geometry manipulation and plotting libraries
vtk>=9.2
pyvista>=0.38
pyvistaqt>=0.9
matplotlib>=3.6
shapely>=2.0
rasterio>=1.3
Pillow>=9.4
